tidy up of Headless.Scripting.CSharp
