--- Headless.CSharp/net8/.using.cs
@@ -1,14 +1,0 @@
-﻿#if NET8_0_OR_GREATER
-global using Microsoft.CodeAnalysis;
-global using Microsoft.CodeAnalysis.CSharp;
-global using Microsoft.CodeAnalysis.CSharp.Scripting;
-global using Microsoft.CodeAnalysis.CSharp.Syntax;
-global using Microsoft.CodeAnalysis.Scripting;
-global using Headless.CSharp.Extensions;
-global using System;
-global using System.Diagnostics;
-global using System.Collections.Generic;
-global using System.Collections.Immutable;
-global using System.Linq;
-global using System.Reflection;
-#endif
