--- Headless.CSharp/net8/ScriptEntryPointType.cs
@@ -1,11 +1,0 @@
-﻿#if NET8_0_OR_GREATER
-// ReSharper disable once CheckNamespace
-namespace Headless.CSharp;
-
-public enum ScriptEntryPointType
-{
-    Uninitialised,
-    ExpressionBody,
-    MethodBody
-}
-#endif
