+++ Headless.CSharp/net8/Scripting/MethodBodyImplementation.cs
@@ -1,0 +1,9 @@
+﻿// ReSharper disable once CheckNamespace
+namespace Headless.CSharp.Scripting;
+
+public enum MethodBodyImplementation
+{
+    Uninitialised,
+    Block,
+    Expression
+}
