+++ Headless.CSharp/net8/Scripting/MethodParameter.cs
@@ -1,0 +1,4 @@
+﻿// ReSharper disable once CheckNamespace
+namespace Headless.CSharp.Scripting;
+
+public record MethodParameter(int Order, string? Name, Type? Type);
