+++ Headless.CSharp/net8/Scripting/ScriptImplementationDescriptor.cs
@@ -1,0 +1,4 @@
+﻿// ReSharper disable once CheckNamespace
+namespace Headless.CSharp.Scripting;
+
+public record ScriptImplementationDescriptor(MethodBodyImplementation EntryPointBodyType, MethodParameter[] ParameterDescriptors);
